# Conflicts:
#	.gitignore
#	nts.uk/uk.repo/nts.arc.jar
#	nts.uk/uk.repo/nts.gul.jar

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/command/application/lateorleaveearly/CanceLateOrLeaveEarlyCommand.java
@@ -10,7 +10,7 @@ import nts.arc.time.GeneralDate;
  */
 @Getter
 public class CanceLateOrLeaveEarlyCommand {
-	private Long version;
+	private int version;
 	
 	/** 会社ID */
 	private String companyID;
