+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/command/application/lateorleaveearly/DeleteLateOrLeaveEarlyCommand.java
@@ -9,7 +9,7 @@ import lombok.Value;
  */
 @Value
 public class DeleteLateOrLeaveEarlyCommand {
-	private Long version;
+	private int version;
 	private String companyID;
 	private String appID;
 }
